DEVELOPMENT.md: improve release documentation
Previously, the release tarball generation description recommended
tagnames following the <package>-<version> pattern.  Since GitHub
provides direct download facilities for the releases based on the
tags, the filenames obtained from GitHub resulted in archives (and
directories) named <package>-<package>-<version>.tgz

This patch updates the documentation to create tags named after the
<version> only, which allow the direct download of the releases from
GitHub.

As a result, the references to "pristine-tar" are augmented with a note
that this tool is an optional extension.